ai/requirements.txt
# ============================================================
# Dependencies — Training + local inference
# ============================================================

# Data
pandas==2.2.2
numpy==1.26.4
pyarrow==16.0.0
kaggle==1.6.17

# ML
scikit-learn==1.5.0
xgboost==2.0.3
optuna==3.6.1

# AWS
boto3==1.34.100

# Notebooks
jupyter==1.0.0
matplotlib==3.9.0
seaborn==0.13.2

# Utils
python-dotenv==1.0.1
